Fix cartItemCard to be clickable and navigate to detail

+++ src/components/cartItemCard/CartItemCard.tsx
@@ -1,12 +1,13 @@
-import React, { useState } from 'react';
-import { Box, Button, ButtonGroup, Dialog, DialogActions, DialogTitle, Divider, Typography } from '@mui/material';
+import { useState } from 'react';
+import { Box, Link, Button, ButtonGroup, Dialog, DialogActions, DialogTitle, Divider, Typography } from '@mui/material';
 import FavoriteBorderIcon from '@mui/icons-material/FavoriteBorder';
 import DeleteOutlineIcon from '@mui/icons-material/DeleteOutline';
 
 import CartItem from '../../misc/types/CartItem';
 import { useAppDispatch } from '../../redux/store';
-import { removeFromCart } from '../../redux/slices/CartSlicer';
-import Product from '../../misc/types/Product';
+import { removeFromCart, updateQuantityInCart } from '../../redux/slices/CartSlicer';
+import FormSelects from '../ui/FormSelects';
+import { Link as RouterLink } from 'react-router-dom';
 
 type Props = {
   cartItem: CartItem;
@@ -17,15 +18,29 @@ export default function CartItemCard(props: Props) {
   const [showDeleteDialog, setShowDeleteDialog] = useState<boolean>(false);
   const { item, quantity } = props.cartItem;
   
+  let quantityItems: {key: number}[] = [];
+  for (let i = 0; i < 10; i++) {
+    quantityItems[i] = { key: (i+1) };
+  }
+  
   const handleDeleteItem = () => {
-    console.log('delete', item?.title);
     setShowDeleteDialog(true);
   }
 
   const handleClose = (shouldDelete: boolean = false) => {
     console.log('should delete', shouldDelete);
     setShowDeleteDialog(false);
-    dispatch(removeFromCart(props.cartItem));
+    if (shouldDelete) {
+      dispatch(removeFromCart(props.cartItem));
+    }
+  }
+
+  const handleQuantityChanges = (value: string) => {
+    console.log('handleQuantityChanges', value);
+    dispatch(updateQuantityInCart({
+      ...props.cartItem,
+      quantity: parseInt(value)
+    }));
   }
 
   return (
@@ -33,9 +48,11 @@ export default function CartItemCard(props: Props) {
     <Divider />
     <Box component={'div'} display={'flex'} my={1}>
       <Box component={'div'} width={164}>
-        <img src={item.images && item.images[0] ? item.images[0] :''}
-             width={164}
-             alt={item.title} />
+        <Link href={`/product/${item.id}`} underline="none" color={'inherit'}>
+          <img src={item.images && item.images[0] ? item.images[0] :''}
+                width={164}
+                alt={item.title} />
+        </Link>
       </Box>
       <Box component={'div'} marginLeft={2} width={'100%'}
             display={'flex'} flexWrap={'wrap'} alignContent={'space-between'}>
@@ -51,9 +68,16 @@ export default function CartItemCard(props: Props) {
           </Typography>
         </Box>
         <Box display={'flex'} justifyContent={'space-between'} width={'100%'} alignItems={'center'}>
-          <Box>
+          <Box display={'flex'} alignItems={'center'}>
             {/* if changed, dispatch the change to redux */}
-            Quantity: {quantity}
+            Quantity:
+            <FormSelects 
+              items={quantityItems}  
+              displayKey={'key'} 
+              valueKey={'key'} 
+              size={'small'}
+              selectedValue={quantity.toString()}
+              onChange={handleQuantityChanges} />
           </Box>
 
           <ButtonGroup variant="text" aria-label="Basic button group">
@@ -68,14 +92,15 @@ export default function CartItemCard(props: Props) {
         onClose={() => handleClose(false)}
         aria-labelledby="alert-dialog-title"
         aria-describedby="alert-dialog-description">
-        <DialogTitle id="alert-dialog-title" fontSize={18}>
-          Delete
+        <DialogTitle id="alert-dialog-title" fontSize={15}>
+          Remove
           <span style={{ fontWeight: 'bold'}}> {item.title} </span> from cart?
         </DialogTitle>
         <DialogActions>
           <Button 
             size={'small'}       
             variant={"text"} 
+            sx={{ fontSize: 12}}
             onClick={() => handleClose(false)}>
             Cancel
           </Button>
@@ -83,7 +108,7 @@ export default function CartItemCard(props: Props) {
             size={'small'} 
             variant="contained" 
             color="error"
-            sx={{ borderRadius: 15 }} 
+            sx={{ borderRadius: 15, fontSize: 12 }} 
             onClick={() => handleClose(true)} autoFocus>
             Delete
           </Button>
